+++ internal/packetfabric/billing.go
@@ -80,10 +80,19 @@ func (c *PFClient) ModifyBilling(cID string, billing BillingUpgrade) (*BillingUp
 
 func (c *PFClient) GetEarlyTerminationLiability(circuitID string) (float64, error) {
 	formattedURI := fmt.Sprintf(etlURI, circuitID)
+
 	var resp float64
-	_, err := c.sendRequest(formattedURI, getMethod, nil, &resp)
+	getETL := func() (interface{}, error) {
+		return c.sendRequest(formattedURI, getMethod, nil, &resp)
+	}
+
+	_, err := c.Retry(fmt.Sprintf("get etl for %s", circuitID), getETL)
 	if err != nil {
+		if c.Is404(err) {
+			return 0, nil
+		}
 		return 0, err
 	}
+
 	return resp, nil
 }