+++ internal/packetfabric/cloud_service_google.go
@@ -2,6 +2,7 @@ package packetfabric
 
 import (
 	"fmt"
+	"github.com/go-playground/validator/v10"
 	"time"
 )
 
@@ -34,6 +35,7 @@ type GoogleCloudRouterConn struct {
 	Speed                    string         `json:"speed,omitempty"`
 	PublishedQuoteLineUUID   string         `json:"published_quote_line_uuid,omitempty"`
 	PONumber                 string         `json:"po_number,omitempty"`
+	SubscriptionTerm         int            `json:"subscription_term,omitempty" validate:"oneof=1 12 24 36" default:"1"`
 	CloudSettings            *CloudSettings `json:"cloud_settings,omitempty"`
 }
 
@@ -95,6 +97,9 @@ func (c *PFClient) CreateRequestHostedGoogleMktConn(googleConn GoogleMktCloudCon
 }
 
 func (c *PFClient) CreateGoogleCloudRouterConn(googleConn GoogleCloudRouterConn, cID string) (*CloudRouterConnectionReadResponse, error) {
+	if err := validator.New().Struct(googleConn); err != nil {
+		return nil, err
+	}
 	formatedURI := fmt.Sprintf(servicesGoogleCloudRouterConnURI, cID)
 	expectedResp := &CloudRouterConnectionReadResponse{}
 	_, err := c.sendRequest(formatedURI, postMethod, googleConn, expectedResp)