+++ internal/provider/resource_outbound_cross_connect.go
@@ -2,10 +2,10 @@ package provider
 
 import (
 	"context"
+	"fmt"
 	"time"
 
 	"github.com/PacketFabric/terraform-provider-packetfabric/internal/packetfabric"
-	"github.com/google/uuid"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
@@ -111,6 +111,41 @@ func resourceOutboundCrossConnect() *schema.Resource {
 				ValidateFunc: validation.StringIsNotEmpty,
 				Description:  "The user description used for update.",
 			},
+			"outbound_cross_connect_id": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "The Outbound Cross Connect Outbound Cross Connect ID.",
+			},
+			"obcc_status": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "The Outbound Cross Connect OBCC Status.",
+			},
+			"progress": {
+				Type:        schema.TypeInt,
+				Computed:    true,
+				Description: "The Outbound Cross Connect Progress.",
+			},
+			"deleted": {
+				Type:        schema.TypeBool,
+				Computed:    true,
+				Description: "The Outbound Cross Connect delete state.",
+			},
+			"z_loc_cfa": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "The Outbound Cross Connect Panel/module/position.",
+			},
+			"time_created": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "The Outbound Cross Connect Time created.",
+			},
+			"time_updated": {
+				Type:        schema.TypeString,
+				Computed:    true,
+				Description: "The Outbound Cross Connect Time updated.",
+			},
 		},
 		Importer: &schema.ResourceImporter{
 			StateContext: schema.ImportStatePassthroughContext,
@@ -127,25 +162,73 @@ func resourceOutboundCrossConnectCreate(ctx context.Context, d *schema.ResourceD
 	if err != nil {
 		return diag.FromErr(err)
 	}
-	d.SetId(uuid.New().String())
+
+	circuitID, err := getCrossConnectCircuitID(c, d.Get("port").(string))
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
+	if circuitID == "" {
+		return diag.Errorf("Failed to find the cross connect with port: %s", d.Get("port").(string))
+	}
+
+	d.SetId(circuitID)
+
 	return diags
+}
+
+func getCrossConnectCircuitID(c *packetfabric.PFClient, port string) (string, error) {
+	timeout := time.After(30 * time.Second)
+	ticker := time.NewTicker(2 * time.Second)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ticker.C:
+			crossConns, err := c.ListOutboundCrossConnects()
+			if err != nil {
+				return "", err
+			}
+
+			for _, crossConn := range *crossConns {
+				if crossConn.Port == port {
+					return crossConn.CircuitID, nil
+				}
+			}
+		case <-timeout:
+			return "", fmt.Errorf("timed out waiting for cross connect with port: %s", port)
+		}
+	}
 }
 
 func resourceOutboundCrossConnectRead(ctx context.Context, d *schema.ResourceData, m interface{}) diag.Diagnostics {
 	c := m.(*packetfabric.PFClient)
 	c.Ctx = ctx
 	var diags diag.Diagnostics
-	if crossConnID, ok := d.GetOk("data_center_cross_connect_id"); ok {
-		resp, err := c.GetOutboundCrossConnect(crossConnID.(string))
-		if err != nil {
-			return diag.FromErr(err)
-		}
-		diags = append(diags, diag.Diagnostic{
-			Severity: diag.Warning,
-			Summary:  "Existing outbound cross connect ID",
-			Detail:   resp.OutboundCrossConnectID,
-		})
+
+	resp, err := c.GetOutboundCrossConnect(d.Id())
+	if err != nil {
+		return diag.FromErr(err)
 	}
+	_ = d.Set("port", resp.Port)
+	_ = d.Set("site", resp.Site)
+	_ = d.Set("document_uuid", resp.DocumentUUID)
+	_ = d.Set("outbound_cross_connect_id", resp.OutboundCrossConnectID)
+	_ = d.Set("obcc_status", resp.ObccStatus)
+	_ = d.Set("description", resp.Description)
+	_ = d.Set("user_description", resp.UserDescription)
+	_ = d.Set("destination_name", resp.DestinationName)
+	_ = d.Set("destination_circuit_id", resp.DestinationCircuitID)
+	_ = d.Set("panel", resp.Panel)
+	_ = d.Set("module", resp.Module)
+	_ = d.Set("position", resp.Position)
+	_ = d.Set("data_center_cross_connect_id", resp.DataCenterCrossConnectID)
+	_ = d.Set("progress", resp.Progress)
+	_ = d.Set("deleted", resp.Deleted)
+	_ = d.Set("z_loc_cfa", resp.ZLocCfa)
+	_ = d.Set("time_created", resp.TimeCreated)
+	_ = d.Set("time_updated", resp.TimeUpdated)
+
 	return diags
 }
 
@@ -172,11 +255,9 @@ func resourceOutboundCrossConnectDelete(ctx context.Context, d *schema.ResourceD
 	c := m.(*packetfabric.PFClient)
 	c.Ctx = ctx
 	var diags diag.Diagnostics
-	if port, ok := d.GetOk("port"); ok {
-		err := c.DeleteOutboundCrossConnect(port.(string))
-		if err != nil {
-			return diag.FromErr(err)
-		}
+	err := c.DeleteOutboundCrossConnect(d.Id())
+	if err != nil {
+		return diag.FromErr(err)
 	}
 	return diags
 }