+++ internal/testutil/test_hcls.go
@@ -121,7 +121,7 @@ const GoogleNetwork = "default"
 // packetfabric_cs_ibm_hosted_connection
 const IbmBgpAsn = 64536
 const IbmRegion = "us-east"
-const IbmSpeed = 100 // must match const CloudRouterConnSpeed and HostedCloudSpeed
+const IbmSpeed = 50
 
 // packetfabric_cloud_router_connection_oracle
 // packetfabric_cs_oracle_hosted_connection
@@ -346,13 +346,14 @@ type RHclCloudRouterInput struct {
 // packetfabric_cloud_router_connection_aws
 type RHclCloudRouterConnectionAwsResult struct {
 	HclResultBase
-	CloudRouter  RHclCloudRouterResult
-	AwsAccountID string
-	AccountUuid  string
-	Desc         string
-	Pop          string
-	Zone         string
-	Speed        string
+	CloudRouter      RHclCloudRouterResult
+	AwsAccountID     string
+	AccountUuid      string
+	Desc             string
+	Pop              string
+	Zone             string
+	Speed            string
+	SubscriptionTerm int
 }
 
 // packetfabric_cloud_router_connection_google
@@ -842,21 +843,47 @@ func RHclOutboundCrossConnect() RHcloutboundCrossConnectsResult {
 	}
 }
 
+func LinkAggregationGroupPort() RHclPortResult {
+	resourceName, hclName := GenerateUniqueResourceName(pfPort)
+	uniqueDesc := GenerateUniqueName()
+
+	media := "LX"
+	pop := "LAB1"
+	zone := "D"
+	speed := "1Gbps"
+	portEnabled := false
+
+	hcl := fmt.Sprintf(
+		RResourcePort,
+		hclName,
+		uniqueDesc,
+		media,
+		pop,
+		zone,
+		speed,
+		subscriptionTerm,
+		portEnabled,
+	)
+	return RHclPortResult{
+		HclResultBase: HclResultBase{
+			Hcl:          hcl,
+			Resource:     pfPort,
+			ResourceName: hclName,
+		},
+		ResourceName:     resourceName,
+		Description:      uniqueDesc,
+		Media:            media,
+		Pop:              pop,
+		Zone:             zone,
+		Speed:            speed,
+		SubscriptionTerm: subscriptionTerm,
+		Enabled:          portEnabled,
+	}
+}
+
 // packetfabric_link_aggregation_group
 func RHclLinkAggregationGroup() RHclLinkAggregationGroupResult {
-
-	c, err := _createPFClient()
-	if err != nil {
-		log.Panic(err)
-	}
-
-	portDetails1 := PortDetails{
-		PFClient:     c,
-		DesiredSpeed: portSpeed,
-	}
-
-	hclPortResult1 := portDetails1.RHclPort(false)
-
+	portHclResult := LinkAggregationGroupPort()
 	resourceName, hclName := GenerateUniqueResourceName(pfLinkAggregationGroup)
 	uniqueDesc := GenerateUniqueName()
 	log.Printf("Resource: %s, Resource name: %s, description: %s\n", pfLinkAggregationGroup, hclName, uniqueDesc)
@@ -865,12 +892,11 @@ func RHclLinkAggregationGroup() RHclLinkAggregationGroupResult {
 		hclName,
 		uniqueDesc,
 		LinkAggGroupInterval,
-		hclPortResult1.ResourceName,
-		hclPortResult1.Pop,
-		resourceName,
+		portHclResult.ResourceName,
+		portHclResult.Pop,
 	)
 
-	hcl := fmt.Sprintf("%s\n%s", hclPortResult1.Hcl, linkAggGroupHcl)
+	hcl := fmt.Sprintf("%s\n%s", portHclResult.Hcl, linkAggGroupHcl)
 
 	return RHclLinkAggregationGroupResult{
 		HclResultBase: HclResultBase{
@@ -881,9 +907,9 @@ func RHclLinkAggregationGroup() RHclLinkAggregationGroupResult {
 		Desc:     uniqueDesc,
 		Interval: LinkAggGroupInterval,
 		Members: []string{
-			hclPortResult1.ResourceName,
+			portHclResult.ResourceName,
 		},
-		Pop: hclPortResult1.Pop,
+		Pop: portHclResult.Pop,
 	}
 }
 
@@ -1042,7 +1068,8 @@ func RHclCloudRouter(input RHclCloudRouterInput) RHclCloudRouterResult {
 		CloudRouterASN,
 		input.Capacity,
 		CloudRouterRegionUS,
-		CloudRouterRegionUK)
+		CloudRouterRegionUK,
+		subscriptionTerm)
 
 	return RHclCloudRouterResult{
 		HclResultBase: HclResultBase{
@@ -1087,6 +1114,7 @@ func RHclCloudRouterConnectionAws() RHclCloudRouterConnectionAwsResult {
 		uniqueDesc,
 		pop,
 		zone,
+		subscriptionTerm,
 		CloudRouterConnSpeed)
 
 	hcl := fmt.Sprintf("%s\n%s", hclCloudRouterRes.Hcl, crcHcl)
@@ -1098,11 +1126,12 @@ func RHclCloudRouterConnectionAws() RHclCloudRouterConnectionAwsResult {
 			ResourceName:           resourceName,
 			AdditionalResourceName: hclCloudRouterRes.ResourceName,
 		},
-		AwsAccountID: os.Getenv("PF_AWS_ACCOUNT_ID"),
-		AccountUuid:  os.Getenv("PF_ACCOUNT_ID"),
-		Speed:        CloudRouterConnSpeed,
-		Pop:          pop,
-		Zone:         zone,
+		AwsAccountID:     os.Getenv("PF_AWS_ACCOUNT_ID"),
+		AccountUuid:      os.Getenv("PF_ACCOUNT_ID"),
+		Speed:            CloudRouterConnSpeed,
+		Pop:              pop,
+		Zone:             zone,
+		SubscriptionTerm: subscriptionTerm,
 	}
 }
 
@@ -1149,7 +1178,8 @@ func RHclCloudRouterConnectionGoogle() RHclCloudRouterConnectionGoogleResult {
 		os.Getenv("PF_ACCOUNT_ID"),
 		uniqueDesc,
 		pop,
-		CloudRouterConnSpeed)
+		CloudRouterConnSpeed,
+		subscriptionTerm)
 
 	hcl := fmt.Sprintf("%s\n%s", hclCloudRouterRes.Hcl, crcHcl)
 
@@ -1176,10 +1206,12 @@ func RHclCloudRouterConnectionAzure() RHclCloudRouterConnectionAzureResult {
 	log.Printf("Resource: %s, Resource name: %s, description: %s\n", pfCloudRouterConnAzure, hclName, uniqueDesc)
 
 	host := os.Getenv("PF_HOST")
+	testName := "t1"
 	AzureLocation, AzurePeeringLocation, AzureServiceProviderName := setAzureLocations(host)
 
 	crcHcl := fmt.Sprintf(
 		RResourceCloudRouterConnectionAzure,
+		testName,
 		AzureLocation,
 		AzureLocation,
 		AzureVnetCidr,
@@ -1195,7 +1227,8 @@ func RHclCloudRouterConnectionAzure() RHclCloudRouterConnectionAzureResult {
 		os.Getenv("PF_ACCOUNT_ID"),
 		uniqueDesc,
 		CloudRouterConnSpeed,
-		false)
+		false,
+		subscriptionTerm)
 
 	hcl := fmt.Sprintf("%s\n%s", hclCloudRouterRes.Hcl, crcHcl)
 
@@ -1220,10 +1253,12 @@ func RHclCloudRouterConnectionAzurePublic() RHclCloudRouterConnectionAzureResult
 	log.Printf("Resource: %s, Resource name: %s, description: %s\n", pfCloudRouterConnAzure, hclName, uniqueDesc)
 
 	host := os.Getenv("PF_HOST")
+	testName := "t2"
 	AzureLocation, AzurePeeringLocation, AzureServiceProviderName := setAzureLocations(host)
 
 	crcHcl := fmt.Sprintf(
 		RResourceCloudRouterConnectionAzure,
+		testName,
 		AzureLocation,
 		AzureLocation,
 		AzureVnetCidr,
@@ -1239,7 +1274,8 @@ func RHclCloudRouterConnectionAzurePublic() RHclCloudRouterConnectionAzureResult
 		os.Getenv("PF_ACCOUNT_ID"),
 		uniqueDesc,
 		CloudRouterConnSpeed,
-		true)
+		true,
+		subscriptionTerm)
 
 	hcl := fmt.Sprintf("%s\n%s", hclCloudRouterRes.Hcl, crcHcl)
 
@@ -1287,8 +1323,9 @@ func RHclCloudRouterConnectionIbm() RHclCloudRouterConnectionIbmResult {
 		uniqueDesc,
 		pop,
 		zone,
-		CloudRouterConnSpeed,
+		backboneVCspeed,
 		IbmBgpAsn,
+		subscriptionTerm,
 		IbmRegion,
 		uniqueDesc,
 		IbmBgpAsn,
@@ -1307,7 +1344,7 @@ func RHclCloudRouterConnectionIbm() RHclCloudRouterConnectionIbmResult {
 		AccountUuid: os.Getenv("PF_ACCOUNT_ID"),
 		Pop:         pop,
 		Zone:        zone,
-		Speed:       CloudRouterConnSpeed,
+		Speed:       backboneVCspeed,
 		IbmBgpAsn:   IbmBgpAsn,
 	}
 }
@@ -1328,7 +1365,7 @@ func RHclCloudRouterConnectionOracle() RHclCloudRouterConnectionOracleResult {
 		IsCloudConnection:     true,
 	}
 
-	pop, zone, region := popDetails.FindAvailableCloudPopZone()
+	pop, zone, _ := popDetails.FindAvailableCloudPopZone()
 
 	hclCloudRouterRes := RHclCloudRouter(DefaultRHclCloudRouterInput())
 	resourceName, hclName := GenerateUniqueResourceName(pfCloudRouterConnOracle)
@@ -1337,9 +1374,7 @@ func RHclCloudRouterConnectionOracle() RHclCloudRouterConnectionOracleResult {
 
 	crcHcl := fmt.Sprintf(
 		RResourceCloudRouterConnectionOracle,
-		region,
 		OracleProviderName,
-		region,
 		OracleBandwidth,
 		OracleBgpAsn,
 		OracleAuthKey,
@@ -1351,7 +1386,7 @@ func RHclCloudRouterConnectionOracle() RHclCloudRouterConnectionOracleResult {
 		uniqueDesc,
 		pop,
 		zone,
-		region)
+		subscriptionTerm)
 
 	hcl := fmt.Sprintf("%s\n%s", hclCloudRouterRes.Hcl, crcHcl)
 
@@ -1389,6 +1424,7 @@ func RHclCloudRouterConnectionPort() RHclCloudRouterConnectionPortResult {
 		portTestResult.ResourceName,
 		CloudRouterConnPortSpeed,
 		CloudRouterConnPortVlan,
+		subscriptionTerm,
 	)
 
 	hcl := fmt.Sprintf("%s\n%s\n%s", portTestResult.Hcl, hclCloudRouterRes.Hcl, crConnPortHcl)
@@ -1437,9 +1473,9 @@ func RHclCloudRouterConnectionIpsec() RHclCloudRouterConnectionIpsecResult {
 		CloudRouterConnIpsecPhase1Lifetime,
 		CloudRouterConnIpsecPhase2pfsGroup,
 		CloudRouterConnIpsecPhase2EncryptionAlgo,
-		CloudRouterConnIpsecPhase2AuthenticationAlgo,
 		CloudRouterConnIpsecPhase2Lifetime,
 		CloudRouterConnIpsecSharedKey,
+		subscriptionTerm,
 	)
 
 	hcl := fmt.Sprintf("%s\n%s", hclCloudRouterResult.Hcl, cloudRouterIpsecHcl)
@@ -1462,7 +1498,6 @@ func RHclCloudRouterConnectionIpsec() RHclCloudRouterConnectionIpsecResult {
 		Phase1Lifetime:             CloudRouterConnIpsecPhase1Lifetime,
 		Phase2PfsGroup:             CloudRouterConnIpsecPhase2pfsGroup,
 		Phase2EncryptionAlgo:       CloudRouterConnIpsecPhase2EncryptionAlgo,
-		Phase2AuthenticationAlgo:   CloudRouterConnIpsecPhase2AuthenticationAlgo,
 		Phase2Lifetime:             CloudRouterConnIpsecPhase2Lifetime,
 		SharedKey:                  CloudRouterConnIpsecSharedKey,
 	}
@@ -1470,9 +1505,22 @@ func RHclCloudRouterConnectionIpsec() RHclCloudRouterConnectionIpsecResult {
 
 // packetfabric_cloud_router_bgp_session
 func RHclBgpSession() RHclBgpSessionResult {
+	resourceName, hclName := GenerateUniqueResourceName(pfCloudRouterBgpSession)
+	return RHclBgpSessionDisabled(resourceName, hclName, false)
+}
+
+// packetfabric_cloud_router_bgp_session
+func RHclBgpSessionEnabledAndDisabled() (RHclBgpSessionResult, RHclBgpSessionResult) {
+	resourceName, hclName := GenerateUniqueResourceName(pfCloudRouterBgpSession)
+	enabled := RHclBgpSessionDisabled(resourceName, hclName, false)
+	disabled := RHclBgpSessionDisabled(resourceName, hclName, true)
+	return enabled, disabled
+}
+
+// packetfabric_cloud_router_bgp_session
+func RHclBgpSessionDisabled(resourceName string, hclName string, disabled bool) RHclBgpSessionResult {
 	hclCloudConnRes := RHclCloudRouterConnectionAws()
 
-	resourceName, hclName := GenerateUniqueResourceName(pfCloudRouterBgpSession)
 	log.Printf("Resource: %s, Resource name: %s\n", pfCloudRouterBgpSession, hclName)
 
 	bgpSessionHcl := fmt.Sprintf(
@@ -1480,6 +1528,7 @@ func RHclBgpSession() RHclBgpSessionResult {
 		hclName,
 		hclCloudConnRes.HclResultBase.AdditionalResourceName,
 		hclCloudConnRes.HclResultBase.ResourceName,
+		fmt.Sprintf("%t", disabled),
 		CloudRouterBgpSessionRemoteAddress,
 		CloudRouterBgpSessionL3Address,
 		CloudRouterBgpSessionASN,
@@ -1520,6 +1569,7 @@ func RHclCloudRouterQuickConnect() RHclCloudRouterQuickConnectResult {
 		bgpSessionResult.CloudRouterConn.AdditionalResourceName,
 		bgpSessionResult.CloudRouterConn.ResourceName,
 		os.Getenv("PF_QUICK_CONNECT_SERVICE_UUID"),
+		subscriptionTerm,
 		CloudRouterQCReturnFilterPrefix1,
 		CloudRouterQCReturnFilterType,
 		CloudRouterQCReturnFilterPrefix2,
@@ -1782,7 +1832,7 @@ func RHclCsIbmHostedConnection() RHclCsHostedCloudIbmResult {
 		uniqueDesc,
 		pop,
 		zone,
-		HostedCloudSpeed,
+		backboneVCspeed,
 		HostedCloudVlan4,
 		IbmBgpAsn,
 		IbmRegion,
@@ -1802,7 +1852,7 @@ func RHclCsIbmHostedConnection() RHclCsHostedCloudIbmResult {
 		AccountUuid: os.Getenv("PF_ACCOUNT_ID"),
 		Pop:         pop,
 		Zone:        zone,
-		Speed:       HostedCloudSpeed,
+		Speed:       backboneVCspeed,
 		Vlan:        HostedCloudVlan4,
 		IbmBgpAsn:   IbmBgpAsn,
 	}
@@ -1822,7 +1872,10 @@ func RHclCsOracleHostedConnection() RHclCsHostedCloudOracleResult {
 		DesiredConnectionType: "hosted",
 		IsCloudConnection:     true,
 	}
-	pop, zone, region := popDetails.FindAvailableCloudPopZone()
+	pop, zone, _ := popDetails.FindAvailableCloudPopZone()
+
+	// the region and drg need to match
+	region := os.Getenv("TF_VAR_pf_cs_oracle_region")
 
 	portDetails := CreateBasePortDetails()
 	portTestResult := portDetails.RHclPort(false)
@@ -1886,7 +1939,7 @@ func RHclCsAwsDedicatedConnection() RHclCsAwsDedicatedConnectionResult {
 		IsCloudConnection:     true,
 	}
 	pop, zone, region, err := popDetails.FindAvailableCloudPopZoneDedicated()
-	if err != nil {
+	if err != nil || "" == pop {
 		log.Println("Error getting pop and zone with available port: ", err)
 		log.Panic(err)
 	}